move RMT-CLI behavior in its own class
